+++ jest.config.js
@@ -4,7 +4,11 @@ module.exports = {
 		"<rootDir>/node_modules/",
 		"<rootDir>/src/.*\\.ts",
 	],
-	testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.[j]sx?$',
+	testRegex: '(/__tests__/.*)(test|spec)\\.jsx?$',
 	collectCoverageFrom: ["src/**/*.js"],
+	coveragePathIgnorePatterns: [
+		'__tests__',
+		'/src/test/'
+	],
 	silent: true,
 }