+++ src/classes/__tests__/fixtures/AbstractAppliance.unit.test/ApplianceWithoutInputTypes.ts
@@ -1,0 +1,5 @@
+import { Appliance } from './Appliance'
+
+export class ApplianceWithoutInputTypes extends Appliance {
+	protected static inputTypes = undefined
+}