+++ src/errors/AbstractImplementationError.ts
@@ -1,0 +1,1 @@
+export class AbstractImplementationError extends Error {}