+++ src/errors/ProcessingError.ts
@@ -1,0 +1,1 @@
+export class ProcessingError extends Error {}