+++ src/types/LogLevel.ts
@@ -1,8 +1,0 @@
-export enum LogLevel {
-	Fatal = 'fatal',
-	Error = 'error',
-	Warn = 'warn',
-	Info = 'info',
-	Debug = 'debug',
-	Trace = 'trace',
-}